Resolve last TODO
